Remove solver dir and add solve() to Expression
