+++ src/CMakeLists.txt
@@ -1,2 +1,1 @@
 add_subdirectory(math_objects)
-add_subdirectory(solver)
