--- src/solver/tests/.clang-tidy
@@ -1,1 +1,0 @@
-Checks: '-*'
